// src/types/MonolithCustomerDiscussion.ts
export interface IMonolithCustomerDiscussion {
    link_id: number;
    link_customer: number;
    link_discussion: number;
    link_status: string;
    link_agreed: string;
    link_created: Date;
    link_updated: Date;
    link_uuid: string;
    customer_uuid: string;
    discussion_uuid: string;
}